+++ packages/core/src/Element.tsx
@@ -10,14 +10,8 @@ export interface ElementProps {
 }
 
 export default class Element extends React.PureComponent<ElementProps> {
-  static defaultProps = {
-    attributes: {},
-    children: null,
-    selfClose: false,
-  };
-
   render() {
-    const { attributes, children, selfClose, tagName: Tag } = this.props;
+    const { attributes = {}, children = null, selfClose = false, tagName: Tag } = this.props;
 
     // @ts-ignore BUG: https://github.com/Microsoft/TypeScript/issues/28806
     return selfClose ? <Tag {...attributes} /> : <Tag {...attributes}>{children}</Tag>;